requirements.txt
# ✅ Core dependencies
fastapi==0.103.1
uvicorn==0.23.2
pydantic==1.10.8
streamlit==1.30.0

# ✅ Async requests for FastAPI
aiohttp==3.9.1

# ✅ NLP & Search (BM25 + FAISS)
faiss-cpu==1.7.4
rank-bm25==0.2.2
sentence-transformers==2.2.2

# ✅ Hugging Face API for embeddings & LLM
transformers==4.35.2
torch==2.0.1
langchain==0.2.2
langchain-community==0.2.2
langchain-core==0.2.2
langchain-huggingface==0.2.2

# # ✅ Optional: If using RAGAS for evaluation
# ragas==0.1.0

# ✅ Other utilities
numpy==1.26.0
pandas==2.1.4
